# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cloud IDE API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fs/read:
    get:
      operationId: readFile
      tags: [filesystem]
      summary: Read file content
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
        "404":
          description: File not found
  /fs/write:
    post:
      operationId: writeFile
      tags: [filesystem]
      summary: Write or update file content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileWrite"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    FileContent:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
      required:
        - path
        - content
    FileWrite:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
      required:
        - path
        - content
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required:
        - success
...
    DeployConfig:
      type: object
      properties:
        type:
          type: string
          enum: [static, node, python]
        buildCommand:
          type: string
        startCommand:
          type: string
        port:
          type: integer
        entryPoint:
          type: string
      required:
        - type
        - buildCommand
        - startCommand
        - port
        - entryPoint
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        ownerId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        deployConfig:
          $ref: "#/components/schemas/DeployConfig"
      required:
        - id
        - name
        - description
        - ownerId
        - createdAt
        - updatedAt
        - deployConfig
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        ownerId:
          type: ["string", "null"]
      required:
        - name
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        deployConfig:
          type: ["object", "null"]
...